+++ Sources/SatochipSwift/StatusWord.swift
@@ -31,7 +31,9 @@ public enum StatusWord: UInt16, Error {
     case resetToFactory = 0xFF00 // Card has been reset to factory
     case insDeprecated = 0x9C26 // For instructions that have been deprecated
     case debugFlag = 0x9FFF // for debugging purpose 2
-
+    case wrongLength = 0x6700 // Wrong length error
+    case objectAlreadyPresent = 0x9C60 // Imported object is already present
+    
     // secure channel
     case secureChannelRequired = 0x9C20
     case secureChannelUninitialized = 0x9C21
@@ -55,12 +57,16 @@ public enum StatusWord: UInt16, Error {
     case secureImportNoTrustePubkey = 0x9C35 // No Trusted Pubkey when importing Secret through Secure import
     
     // Satodime
-    
+    case incorrectUnlockCounter = 0x9C50
+    case incorrectUnlockCode = 0x9C51
+    case incorrectKeyslotState = 0x9C52
+    case incorrectProtocolMedia = 0x9C53
+    case unknownProtocolMedia = 0x9C54
+
     // seedkeeeper
     case lockError = 0x9C30
     case exportNotAllowed = 0x9C31
     case usageNotAllowed = 0x9C36
     case wrongSecretType = 0x9C38
-    
 }